+++ subscriber.py
@@ -1,10 +1,9 @@
 import time
-from provenance.capture import Logger
-from provenance.amqp import Subscriber
+from provenance.capture import Logger, Subscriber
 
 
 logger = Logger.stdout()
 
-subscriber = Subscriber('localhost', 'provenance', logger)
+subscriber = Subscriber(logger, 'localhost', 'guest', 'guest', 'provenance')
 
 subscriber.start()
